Move MapFragment to components/map folder
