+++ client/app/src/main/java/enchantedtowers/client/MapActivity.java
@@ -1,9 +1,7 @@
 package enchantedtowers.client;
 
 import android.Manifest;
-import android.content.Intent;
 import android.os.Bundle;
-import android.provider.Settings;
 import android.widget.Toast;
 
 import androidx.activity.result.ActivityResultLauncher;
@@ -12,6 +10,7 @@ import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 
+import enchantedtowers.client.components.map.MapFragment;
 import enchantedtowers.client.components.permissions.PermissionManager;
