+++ client/app/src/main/res/layout/fragment_map.xml
@@ -5,7 +5,7 @@
     xmlns:map="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MapFragment">
+    tools:context=".components.map.MapFragment">
 
     <fragment
         android:layout_width="match_parent"
